+++ packages/cli/src/modules/structurelint/types.ts
@@ -1,0 +1,3 @@
+export interface IModuleStructurelintConfig {
+  path: string;
+}